+++ src/command/factory.ts
@@ -226,11 +226,11 @@ export class CommandFactory {
       case 'COMMAND_PALETTE_CLOSE':
         return new CommandPaletteCloseCommand(this.commandPaletteViewModel);
       case 'DESCRIBE_X':
-        return new DescribeXCommand(this.context, this.textViewModel);
+        return new DescribeXCommand(this.context, this.textViewModel, this.audioService, this.textService);
       case 'DESCRIBE_Y':
-        return new DescribeYCommand(this.context, this.textViewModel);
+        return new DescribeYCommand(this.context, this.textViewModel, this.audioService, this.textService);
       case 'DESCRIBE_FILL':
-        return new DescribeFillCommand(this.context, this.textViewModel);
+        return new DescribeFillCommand(this.context, this.textViewModel, this.audioService, this.textService);
       case 'DESCRIBE_POINT':
         return new DescribePointCommand(
           this.context,
@@ -238,13 +238,14 @@ export class CommandFactory {
           this.highlightService,
           this.brailleViewModel,
           this.textViewModel,
+          this.textService,
         );
       case 'DESCRIBE_TITLE':
-        return new DescribeTitleCommand(this.context, this.textViewModel);
+        return new DescribeTitleCommand(this.context, this.textViewModel, this.audioService, this.textService);
       case 'DESCRIBE_SUBTITLE':
-        return new DescribeSubtitleCommand(this.context, this.textViewModel);
+        return new DescribeSubtitleCommand(this.context, this.textViewModel, this.audioService, this.textService);
       case 'DESCRIBE_CAPTION':
-        return new DescribeCaptionCommand(this.context, this.textViewModel);
+        return new DescribeCaptionCommand(this.context, this.textViewModel, this.audioService, this.textService);
       case 'ANNOUNCE_POSITION':
         return new AnnouncePositionCommand(
           this.context,